keep performance merge logic on native
